remove redundant or empty chap texts file

+++ tasks/routine/clean_zhtexts.cr
@@ -1,8 +1,11 @@
 require "colorize"
+require "../../src/filedb/stores/zip_store"
+
 DIR = "_db/nvdata/zhtexts"
 
 seeds = Dir.children(DIR).each do |seed|
   folders = Dir.glob("#{DIR}/#{seed}/*/")
+
   puts "#{seed}: #{folders.size} folders".colorize.cyan.bold
 
   folders.each_with_index do |folder, idx|
@@ -10,8 +13,12 @@ seeds = Dir.children(DIR).each do |seed|
       puts "- <#{idx + 1}/#{folders.size}> #{folder}".colorize.blue
     end
 
-    Dir.glob(folder + "*.txt").each do |file|
-      next if File.size(file) > 5
+    folder = folder.sub(/\/$/, "") # remove trailing `/`
+    zip_store = CV::ZipStore.new("#{folder}.zip", folder)
+
+    Dir.glob("#{folder}/*.txt").each do |file|
+      next unless File.size(file) < 5 || zip_store.in_zip?(File.basename(file))
+
       puts "-- Delete empty file #{file}".colorize.light_red
       File.delete(file)
     end
